backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
zhipuai>=2.0.1
httpx>=0.27.0
python-multipart>=0.0.9
aiosqlite>=0.20.0
sniffio>=1.3.0
python-docx>=1.1.0
PyPDF2>=3.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
